% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{cacheEnv}
\alias{cacheEnv}
\title{Environment for caching Telraam API responses}
\format{
An object of class \code{environment} of length 0.
}
\usage{
cacheEnv
}
\description{
Internal environment used to cache API responses and reduce redundant requests within the package.
}
\keyword{internal}
