% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_directions.R
\name{get_cardinal_dirs}
\alias{get_cardinal_dirs}
\title{Get cardinal directions for Telraam segments}
\usage{
get_cardinal_dirs(mysegments)
}
\arguments{
\item{mysegments}{An \code{sf} object containing the road segments.}
}
\value{
A \code{data.frame} with columns \code{lft} and \code{rgt} indicating the cardinal directions for each segment side.
}
\description{
Calculates the cardinal directions (e.g., Northbound, Eastbound) for the left and right sides of Telraam road segments based on their geometry.
}
\examples{
\dontrun{
anysegment <- read_telraam_segments() |>
  dplyr::filter(oidn == 9000003890)
get_cardinal_dirs(anysegment)
}

}
