% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_telraam_token}
\alias{get_telraam_token}
\title{Get the Telraam API authentication token}
\usage{
get_telraam_token()
}
\value{
Character string containing the Telraam API token. Throws an error if not set.
}
\description{
Retrieves the Telraam API token from the environment.
}
\examples{
\dontrun{
# After setting the token with set_telraam_token()
get_telraam_token()
}
}
