% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cameras.R
\name{read_telraam_cameras}
\alias{read_telraam_cameras}
\title{Read all cameras}
\usage{
read_telraam_cameras(mytoken = get_telraam_token())
}
\arguments{
\item{mytoken}{the authentication token, if not previously set with the \code{set_TelraamToken} function}
}
\value{
a data frame with the data for all cameras in telraam
}
\description{
Read all cameras
}
\examples{

\dontrun{
read_telraam_cameras()
}
}
