% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{read_telraam_segments}
\alias{read_telraam_segments}
\title{Obtain the telraam segments as a sf object}
\usage{
read_telraam_segments(mytoken = get_telraam_token())
}
\arguments{
\item{mytoken}{the authentication token, if not previously set with \code{usethis::edit_r_environ()} or the \code{set_TelraamToken} function}
}
\value{
a sf object with all segments with the id
}
\description{
Obtain the telraam segments as a sf object
}
\examples{
\dontrun{
read_telraam_segments()
}

}
