% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_telraam_token}
\alias{set_telraam_token}
\title{Set the Telraam API authentication token}
\usage{
set_telraam_token(token)
}
\arguments{
\item{token}{Character string. The Telraam API token to be saved.}
}
\value{
Logical TRUE if the token is correctly set, otherwise an error is thrown.
}
\description{
Stores the provided authentication token in the environment for use with Telraam API requests.
}
\examples{
mytoken <- "ivRgw7ZAGFedfwIdASezecdnETZDsdETB4Bqv3pbs5X8JDNnt1pQtpxDmpR6as2k"
\dontrun{
set_telraam_token(mytoken)
}
}
