% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_directions.R
\name{get_cardinal_dirs}
\alias{get_cardinal_dirs}
\title{Get the cardinal directions for the segments}
\usage{
get_cardinal_dirs(mysegments)
}
\arguments{
\item{mysegments}{an \code{sf} object with the segments}
}
\value{
a \code{data.frame} with the correspondence of left \code{lft} and right \code{rgt}directions
to cardinal directions of the road
}
\description{
Get the cardinal directions for the segments
}
\examples{
\dontrun{
anysegment = my_segments |>
                filter(oidn == 9000003890)

get_cardinal_dirs(anysegment)
}
}
