% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_data.R
\name{read_Telraam_traffic}
\alias{read_Telraam_traffic}
\title{Get hourly traffic report}
\usage{
read_Telraam_traffic(
  id,
  report = c("per-hour"),
  time_start,
  time_end,
  tz = Sys.timezone(),
  mytoken = Sys.getenv("telraam")
)
}
\arguments{
\item{id}{the segment (or instance) identifier in question (can be found in the address of the segment from the Telraam website}

\item{report}{can only be "per-hour", resulting in hourly aggregated traffic}

\item{time_start}{The beginning of the requested time interval}

\item{time_end}{The end of the requested time interval (note: the time interval is closed-open, so the end time is not included anymore in the request}

\item{tz}{timezone, by default the system timezone}

\item{mytoken}{the authentication token, if not previously set with the \code{set_TelraamToken} function}
}
\value{
data.frame with traffic data
}
\description{
Function as an interface for the traffic API call
}
\examples{
read_Telraam_traffic(9000003890,time_start = "2023-03-28 07:00:00",time_end = "2023-03-30 07:00:00")
}
