% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_Telraam_Token}
\alias{set_Telraam_Token}
\title{Saves an Authentication Token for the telraam API}
\usage{
set_Telraam_Token(token)
}
\arguments{
\item{token}{a \code{string} with the token}
}
\value{
TRUE if token correctly set
}
\description{
Saves an Authentication Token for the telraam API
}
\examples{

mytoken = "ivRgw7ZAGFedfwIdASezecdnETZDsdETB4Bqv3pbs5X8JDNnt1pQtpxDmpR6as2k"
set_Telraam_Token(mytoken)

}
